fixes #73. Cleanups to UNSET treatment, particularly for for-loop iterants
